refactor(clawdstrike-logos): extract load_time from verifier.rs
Move load-time verification orchestration (verify_policy_at_load_time and
its resolver/parent variants, the inner pipeline, receipt enrichment) plus
the LRU VerificationCache, its private VerificationCacheState, the registration
statics, and install_clawdstrike_policy_load_verifier into
verifier/load_time.rs. The verifier.rs root is now ~63 lines of doc, imports,
module wiring, explicit public re-exports, and the lone cfg(test) helper (the
inline mod tests block is extracted in the next step). The load_time glob is
reached only from the test module, so it carries #[allow(unused_imports)]
mirroring the api_server precedent. The single z3-gated load_time_verifier
branch is preserved verbatim. No logic changes.
